+++ src/app/[locale]/viewer/_components/DummyComponent.tsx
@@ -3,11 +3,12 @@
 import { Box } from '@mui/material';
 import { useTranslations } from 'next-intl';
 
-
 export function DummyComponent() {
     const t = useTranslations('submodels.timeSeries');
 
     return (
-        <Box>{t('linkedSegments')}</Box>
+        <Box>
+            {t('linkedSegments')}
+        </Box>
     )
 }
